Restructuration appli (fichiers séparés dans front/src/)
