+++ front/src/components/jeuxHtml.js
@@ -1,0 +1,12 @@
+// intro splash
+const jeuxHtml = `
+<div id="galaxy">
+  <div class="bg"></div>
+  <div class="stars-back"></div>
+  <div class="stars-middle"></div>
+  <div class="stars-front"></div>
+  <div class="bg center"></div>
+</div>
+<div id="quiz"></div>`
+
+module.exports = jeuxHtml
