+++ front/src/components/splashScreenHtml.js
@@ -1,0 +1,9 @@
+const  splashScreenHtml  = `
+<div class="starwars-demo">
+          <img src="https://sylvainkosc.github.io/planete/1love.png" alt="Star" class="star"/>
+          <img src="https://sylvainkosc.github.io/planete/nebuleuse.png" alt="Wars" class="wars"/>
+    <h2 class="byline" id="byline">Watch out for the black hole</h2>
+</div>
+`
+
+module.exports = splashScreenHtml
