+++ public/app.js
@@ -1,3 +1,138 @@
+/******/ (function(modules) { // webpackBootstrap
+/******/ 	// The module cache
+/******/ 	var installedModules = {};
+/******/
+/******/ 	// The require function
+/******/ 	function __webpack_require__(moduleId) {
+/******/
+/******/ 		// Check if module is in cache
+/******/ 		if(installedModules[moduleId]) {
+/******/ 			return installedModules[moduleId].exports;
+/******/ 		}
+/******/ 		// Create a new module (and put it into the cache)
+/******/ 		var module = installedModules[moduleId] = {
+/******/ 			i: moduleId,
+/******/ 			l: false,
+/******/ 			exports: {}
+/******/ 		};
+/******/
+/******/ 		// Execute the module function
+/******/ 		modules[moduleId].call(module.exports, module, module.exports, __webpack_require__);
+/******/
+/******/ 		// Flag the module as loaded
+/******/ 		module.l = true;
+/******/
+/******/ 		// Return the exports of the module
+/******/ 		return module.exports;
+/******/ 	}
+/******/
+/******/
+/******/ 	// expose the modules object (__webpack_modules__)
+/******/ 	__webpack_require__.m = modules;
+/******/
+/******/ 	// expose the module cache
+/******/ 	__webpack_require__.c = installedModules;
+/******/
+/******/ 	// define getter function for harmony exports
+/******/ 	__webpack_require__.d = function(exports, name, getter) {
+/******/ 		if(!__webpack_require__.o(exports, name)) {
+/******/ 			Object.defineProperty(exports, name, {
+/******/ 				configurable: false,
+/******/ 				enumerable: true,
+/******/ 				get: getter
+/******/ 			});
+/******/ 		}
+/******/ 	};
+/******/
+/******/ 	// getDefaultExport function for compatibility with non-harmony modules
+/******/ 	__webpack_require__.n = function(module) {
+/******/ 		var getter = module && module.__esModule ?
+/******/ 			function getDefault() { return module['default']; } :
+/******/ 			function getModuleExports() { return module; };
+/******/ 		__webpack_require__.d(getter, 'a', getter);
+/******/ 		return getter;
+/******/ 	};
+/******/
+/******/ 	// Object.prototype.hasOwnProperty.call
+/******/ 	__webpack_require__.o = function(object, property) { return Object.prototype.hasOwnProperty.call(object, property); };
+/******/
+/******/ 	// __webpack_public_path__
+/******/ 	__webpack_require__.p = "";
+/******/
+/******/ 	// Load entry module and return exports
+/******/ 	return __webpack_require__(__webpack_require__.s = 0);
+/******/ })
+/************************************************************************/
+/******/ ([
+/* 0 */
+/***/ (function(module, exports, __webpack_require__) {
+
+const splashScreenHtml = __webpack_require__(1)
+const jeuxHtml = __webpack_require__(2)
+const questions = __webpack_require__(3)
+
+const contentContainer = document.getElementById("content")
+
+fetch("https://cdn.rawgit.com/akabab/starwars-api/0.2.1/api/all.json")
+.then(response => response.json())
+.then(data => console.log(data))
+
+function afficheJeux() {
+    contentContainer.innerHTML = jeuxHtml
+    // affiche les questions
+    questions()
+}
+
+// intro splash
+function afficheSplashScreen() {
+  contentContainer.innerHTML = splashScreenHtml;
+
+  setTimeout(
+    afficheJeux,
+    1000
+  )
+}
+afficheSplashScreen()
+// fin intro splash
+
+
+/***/ }),
+/* 1 */
+/***/ (function(module, exports) {
+
+const  splashScreenHtml  = `
+<div class="starwars-demo">
+          <img src="https://sylvainkosc.github.io/planete/1love.png" alt="Star" class="star"/>
+          <img src="https://sylvainkosc.github.io/planete/nebuleuse.png" alt="Wars" class="wars"/>
+    <h2 class="byline" id="byline">Watch out for the black hole</h2>
+</div>
+`
+
+module.exports = splashScreenHtml
+
+
+/***/ }),
+/* 2 */
+/***/ (function(module, exports) {
+
+// intro splash
+const jeuxHtml = `
+<div id="galaxy">
+  <div class="bg"></div>
+  <div class="stars-back"></div>
+  <div class="stars-middle"></div>
+  <div class="stars-front"></div>
+  <div class="bg center"></div>
+</div>
+<div id="quiz"></div>`
+
+module.exports = jeuxHtml
+
+
+/***/ }),
+/* 3 */
+/***/ (function(module, exports) {
+
 module.exports = (function() {
   var questions = [
   {
@@ -251,3 +386,7 @@ console.log(quiz)
     // return score;
   }
 })
+
+
+/***/ })
+/******/ ]);
